Refactored services and controllers to use new user account services
