+++ src/lib/manage.password.js
